Rename command, setup, cleanup; update changelog
The `:real/command`, `:real/setup` and `:real/cleanup` keys have been
renamed to remove the `:real` prefix. I think this is appropriate
because the meaning of the `:real` prefix has changed. (It no longer
means "can interact with the system under test", now it means "is given
actual results of methods, not symbolic values".)